+++ build.gradle
@@ -62,6 +62,7 @@ subprojects {
     }
 
     dependencies {
+        compile 'org.springframework.boot:spring-boot-starter'
         compileOnly 'org.projectlombok:lombok'
         compile 'org.springframework.boot:spring-boot-starter-log4j2'
         annotationProcessor 'org.projectlombok:lombok'
@@ -128,10 +129,7 @@ dependencies {
     compile 'mysql:mysql-connector-java'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
 
-    testImplementation('org.springframework.boot:spring-boot-starter-test') {
-        exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
-    }
-    testImplementation 'io.rest-assured:rest-assured:3.3.0'
+
 }
