+++ devbie-app-api/build.gradle
@@ -11,6 +11,12 @@ dependencies {
     compile group: 'io.springfox', name: 'springfox-swagger-ui', version: '2.9.2'
     compile "org.flywaydb:flyway-core"
 
+    testImplementation('org.springframework.boot:spring-boot-starter-test') {
+        exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
+    }
+//    testImplementation 'io.rest-assured:restassured:3.3.0'
+    testCompile group: 'io.rest-assured', name: 'rest-assured', version: '3.3.0'
+
     // todo 고민되는 것
     implementation 'io.jsonwebtoken:jjwt:0.9.1'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
