+++ devbie-app-api/src/test/java/underdogs/devbie/notice/acceptance/NoticeAcceptanceTest.java
