+++ devbie-app-api/src/test/java/underdogs/devbie/recommendation/controller/QuestionRecommendationControllerTest.java
@@ -4,7 +4,7 @@ import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 import static underdogs.devbie.recommendation.acceptance.RecommendationAcceptanceTest.*;
-import static underdogs.devbie.user.domain.UserTest.*;
+import static underdogs.devbie.user.controller.UserControllerTest.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,8 +12,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
 import underdogs.devbie.MvcTest;
-import underdogs.devbie.auth.controller.interceptor.BearerAuthInterceptor;
-import underdogs.devbie.auth.controller.resolver.LoginUserArgumentResolver;
+import underdogs.devbie.auth.interceptor.BearerAuthInterceptor;
+import underdogs.devbie.auth.resolver.LoginUserArgumentResolver;
+import underdogs.devbie.recommendation.QuestionRecommendationController;
 import underdogs.devbie.recommendation.service.QuestionRecommendationService;
 import underdogs.devbie.user.domain.User;
