+++ devbie-app-api/src/test/resources/application.properties
@@ -1,0 +1,2 @@
+spring.datasource.data=
+spring.flyway.enabled=false
