+++ devbie-app-api/src/test/resources/notice_create.sql
@@ -1,0 +1,24 @@
+INSERT INTO notice(name, image, job_position, title, notice_type, content, recruitment_type) VALUES
+('bossdog2',  'https://cdn.vuetifyjs.com/images/cards/road.jpg', 'BACKEND', 'title', 'JOB','hi!', 'ANY')
+,('bossdog4',  'https://cdn.vuetifyjs.com/images/cards/house.jpg', 'BACKEND', 'title', 'JOB','hi!', 'ANY')
+,('bossdog3',  'test', 'BACKEND', 'title', 'JOB','hi!', 'ANY')
+,('bossdog1',  'https://cdn.vuetifyjs.com/images/cards/plane.jpg', 'BACKEND', 'title', 'JOB','hi!', 'ANY')
+,('bossdog5',  'https://cdn.vuetifyjs.com/images/cards/plane.jpg', 'BACKEND', 'title', 'JOB','hi!', 'ANY')
+,('bossdog',  'https://cdn.vuetifyjs.com/images/cards/house.jpg', 'BACKEND', 'title', 'JOB', 'hi!', 'ANY')
+,('bossdog6',  'https://cdn.vuetifyjs.com/images/cards/road.jpg', 'BACKEND', 'title', 'JOB','hi!', 'ANY')
+,('bossdog7',  'test', 'BACKEND', 'title', 'JOB','hi!', 'ANY')
+,('yeji1',  'test', 'BACKEND', 'title', 'JOB','hi!', 'ANY')
+,('bossdog6',  'https://cdn.vuetifyjs.com/images/cards/road.jpg', 'FRONTEND', 'title', 'JOB','hi!', 'ANY')
+,('bossdog7',  'test', 'FRONTEND', 'title', 'JOB','hi!', 'ANY')
+,('yeji1',  'test', 'FRONTEND', 'title', 'JOB','hi!', 'ANY');
+
+INSERT INTO language(notice_id, languages) values
+(1, 'CPP')
+,(1, 'JAVA')
+,(2, 'CPP')
+,(3, 'CPP')
+,(4, 'CPP')
+,(5, 'CPP')
+,(6, 'CPP')
+,(7, 'CPP')
+,(8, 'CPP');
