+++ devbie-app-api/src/test/resources/notice_delete.sql
@@ -1,0 +1,2 @@
+DELETE FROM language;
+DELETE FROM notice;
